# contracts/openapi/music.yaml
openapi: 3.1.0
...
info:
  title: music-view API
  version: 1.0.0
  description: |
    Narzędzie DJ-a do zarządzania biblioteką muzyczną. Rozdział danych:
    `track_catalog` trzyma fakty o utworze, `library_entry` prywatne dane DJ-a.
servers:
  - url: /
    description: aplikacja za proxy /api
...
paths:
  /api/catalog/tracks:
    get:
      tags: [catalog]
      operationId: searchTracks
      summary: Wyszukiwarka katalogu z filtrami, sortowaniem i stronicowaniem
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: genreFamily, in: query, schema: { $ref: '#/components/schemas/GenreFamily' } }
        - { name: yearMin, in: query, schema: { type: integer, format: int32 } }
        - { name: yearMax, in: query, schema: { type: integer, format: int32 } }
        - { name: durationMinSec, in: query, schema: { type: integer, format: int32 } }
        - { name: durationMaxSec, in: query, schema: { type: integer, format: int32 } }
        - { name: popularityMin, in: query, schema: { type: integer, format: int32 } }
        - { name: explicit, in: query, schema: { type: boolean } }
        - { name: bpmMin, in: query, schema: { type: integer, format: int32 } }
        - { name: bpmMax, in: query, schema: { type: integer, format: int32 } }
        - { name: tempoClass, in: query, schema: { $ref: '#/components/schemas/TempoClass' } }
        - { name: energy, in: query, schema: { type: string } }
        - name: camelot
          in: query
          description: Pozycja koła Camelot, np. `8A`
          schema: { type: string }
        - name: camelotCompatible
          in: query
          description: Rozszerza filtr o tonacje zgodne (±1 i równoległa)
          schema: { type: boolean, default: true }
        - { name: inLibrary, in: query, schema: { type: boolean } }
        - { name: ratingMin, in: query, schema: { type: integer, format: int32 } }
        - { name: tag, in: query, schema: { type: string } }
        - { name: valenceMin, in: query, schema: { type: number } }
        - { name: valenceMax, in: query, schema: { type: number } }
        - { name: instrumentalMin, in: query, schema: { type: number } }
        - { name: livenessMax, in: query, schema: { type: number } }
        - { name: bpmSource, in: query, schema: { $ref: '#/components/schemas/BpmSource' } }
        - { name: missing, in: query, schema: { $ref: '#/components/schemas/MissingGroup' } }
        - { name: sort, in: query, schema: { $ref: '#/components/schemas/CatalogSort' } }
        - { name: direction, in: query, schema: { $ref: '#/components/schemas/SortDirection' } }
        - { name: page, in: query, schema: { type: integer, format: int32, default: 0 } }
        - { name: size, in: query, schema: { type: integer, format: int32, default: 20 } }
      responses:
        '200':
          description: Strona wyniku
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageCatalogRowResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/catalog/tracks/{spotifyId}:
    get:
      tags: [catalog]
      operationId: getTrack
      summary: Pojedynczy utwór z katalogu
      parameters: [{ $ref: '#/components/parameters/SpotifyId' }]
      responses:
        '200':
          description: Utwór
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrackResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/enrich:
    post:
      tags: [enrich]
      operationId: startEnrichment
      summary: Startuje job wzbogacania (asynchronicznie)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnrichRequest' }
      responses:
        '202':
          description: Job wystartował
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExecutionIdResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/enrich/estimate:
    post:
      tags: [enrich]
      operationId: estimateEnrichment
      summary: Ile utworów obejmie zlecenie i ile to będzie kosztowało
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnrichRequest' }
      responses:
        '200':
          description: Szacunek
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnrichmentEstimateResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/enrich/jobs:
    get:
      tags: [enrich]
      operationId: listEnrichJobs
      parameters:
        - { name: limit, in: query, schema: { type: integer, format: int32, default: 20 } }
      responses:
        '200':
          description: Historia wykonań
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EnrichJobResponse' }
  /api/enrich/jobs/{executionId}:
    get:
      tags: [enrich]
      operationId: getEnrichJob
      parameters: [{ $ref: '#/components/parameters/ExecutionId' }]
      responses:
        '200':
          description: Status wykonania
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnrichJobResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/enrich/jobs/{executionId}/failures:
    get:
      tags: [enrich]
      operationId: getEnrichJobFailures
      summary: Utwory pominięte w przebiegu wraz z powodem
      parameters:
        - { $ref: '#/components/parameters/ExecutionId' }
        - { name: limit, in: query, schema: { type: integer, format: int32, default: 200 } }
      responses:
        '200':
          description: Lista pominięć
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EnrichFailureResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/enrich/jobs/{executionId}/restart:
    post:
      tags: [enrich]
      operationId: restartEnrichJob
      summary: Restart nieudanego wykonania — dokańcza od checkpointu
      parameters: [{ $ref: '#/components/parameters/ExecutionId' }]
      responses:
        '202':
          description: Nowe wykonanie wystartowało
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExecutionIdResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
    SpotifyId:
      name: spotifyId
      in: path
      required: true
      schema: { type: string }
...
    ExecutionId:
      name: executionId
      in: path
      required: true
      schema: { type: integer, format: int64 }
  responses:
    BadRequest:
      description: Błąd walidacji
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Zasób nie istnieje
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    GenreFamily:
      type: string
      enum: [LATIN, ROCK, POP, DISCO, DISCO_POLO, ELECTRONIC, HIP_HOP, OTHER]
    TempoClass:
      type: string
      enum: [SLOW, MEDIUM, FAST, VERY_FAST]
    BpmSource:
      type: string
      enum: [MANUAL, ACOUSTICBRAINZ, DEEZER, LLM]
    MissingGroup:
      type: string
      enum: [METADATA, AUDIO, AI, ANY]
    CatalogSort:
      type: string
      enum: [RELEVANCE, TITLE, ARTIST, ALBUM, YEAR, BPM, POPULARITY, DURATION,
             DANCEABILITY, ENERGY, RATING, ADDED_AT]
    SortDirection:
      type: string
      enum: [ASC, DESC]
    EnrichmentScope:
      type: string
      enum: [SINGLE, SELECTED, MISSING, OUTDATED]
    FieldGroup:
      type: string
      enum: [METADATA, AUDIO, AI]
...
    ErrorResponse:
      type: object
      required: [errorCode, message, timestamp, traceId]
      properties:
        errorCode: { type: string, examples: [TRACK_NOT_FOUND] }
        message: { type: string }
        timestamp: { type: string, format: date-time }
        traceId: { type: [string, 'null'] }
    TrackResponse:
      type: object
      required: [spotifyId, title, artist, album, year, durationMs, popularity, explicit, albumImageUrl,
                 isrc, genreFamily, style, bpm, bpmSource, danceability, musicalKey,
                 camelot, tempoClass, energy, lyricsTheme, descriptionPl, confidence,
                 enrichedAt, modelUsed, enrichVersion]
      properties:
        spotifyId: { type: string }
        title: { type: [string, 'null'] }
        artist: { type: [string, 'null'] }
        album: { type: [string, 'null'] }
        year: { type: [integer, 'null'], format: int32 }
        durationMs: { type: [integer, 'null'], format: int32 }
        popularity: { type: [integer, 'null'], format: int32 }
        explicit: { type: [boolean, 'null'] }
        albumImageUrl: { type: [string, 'null'] }
        isrc: { type: [string, 'null'] }
        genreFamily: { type: [string, 'null'] }
        style: { type: [string, 'null'] }
        bpm: { type: [integer, 'null'], format: int32 }
        bpmSource: { type: [string, 'null'] }
        danceability: { type: [number, 'null'] }
        musicalKey: { type: [string, 'null'] }
        camelot:
          type: [string, 'null']
          description: Liczona z `musicalKey` przez backend — nie kolumna w bazie
        tempoClass: { type: [string, 'null'] }
        energy: { type: [string, 'null'] }
        lyricsTheme: { type: [string, 'null'] }
        descriptionPl: { type: [string, 'null'] }
        confidence: { type: [string, 'null'] }
        enrichedAt: { type: [string, 'null'], format: date-time }
        modelUsed: { type: [string, 'null'] }
        enrichVersion: { type: [integer, 'null'], format: int32 }
    TrackLibraryResponse:
      type: object
      required: [rating, customTags, addedAt, source]
      properties:
        rating: { type: [integer, 'null'], format: int32 }
        customTags: { type: [array, 'null'], items: { type: string } }
        addedAt: { type: string, format: date-time }
        source: { type: string }
    CatalogRowResponse:
      type: object
      required: [track, library]
      properties:
        track: { $ref: '#/components/schemas/TrackResponse' }
        library:
          oneOf:
            - { $ref: '#/components/schemas/TrackLibraryResponse' }
            - { type: 'null' }
    PageCatalogRowResponse:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/CatalogRowResponse' }
        page: { type: integer, format: int32 }
        size: { type: integer, format: int32 }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer, format: int32 }
...
    EnrichRequest:
      type: object
      required: [scope, fields]
      properties:
        scope: { $ref: '#/components/schemas/EnrichmentScope' }
        fields:
          type: array
          minItems: 1
          uniqueItems: true
          items: { $ref: '#/components/schemas/FieldGroup' }
        spotifyIds:
          type: [array, 'null']
          items: { type: string }
          description: Wymagane dla SINGLE i SELECTED, zabronione dla MISSING i OUTDATED
    ExecutionIdResponse:
      type: object
      required: [executionId]
      properties:
        executionId: { type: integer, format: int64 }
    EnrichmentEstimateResponse:
      type: object
      required: [trackCount, aiTracks, estimatedCost, limit, withinLimit]
      properties:
        trackCount: { type: integer, format: int64 }
        aiTracks:
          type: integer
          format: int64
          description: Utwory idące do LLM-a — sufit liczy się po nich, nie po przebiegu
        estimatedCost:
          type: [number, 'null']
          description: Pusty, gdy w konfiguracji nie ma stawek providera
        limit: { type: integer, format: int32 }
        withinLimit: { type: boolean }
    EnrichJobResponse:
      type: object
      required: [executionId, jobInstanceId, status, scope, fields, readCount, writeCount, failedCount,
                 startTime, endTime, exitDescription]
      properties:
        executionId: { type: integer, format: int64 }
        jobInstanceId: { type: integer, format: int64 }
        status: { type: string }
        scope: { type: [string, 'null'] }
        fields: { type: [string, 'null'] }
        readCount: { type: integer, format: int64 }
        writeCount: { type: integer, format: int64 }
        failedCount: { type: integer, format: int64 }
        startTime: { type: [string, 'null'], format: date-time }
        endTime: { type: [string, 'null'], format: date-time }
        exitDescription: { type: [string, 'null'] }
    EnrichFailureResponse:
      type: object
      required: [spotifyId, reason, failedAt]
      properties:
        spotifyId: { type: string }
        reason: { type: string }
        failedAt: { type: string, format: date-time }
...