# services/music-service/src/main/java/com/pgoogol/music/api/EnrichController.java
package com.pgoogol.music.api;

import com.pgoogol.music.enrichment.EnrichmentEstimate;
import com.pgoogol.music.enrichment.EnrichmentService;
import com.pgoogol.music.enrichment.MissingFieldsCount;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;
import java.util.Map;

@RestController
@RequestMapping("/api/enrich")
@Tag(name = "Enrichment", description = "Joby wzbogacania Spring Batch")
public class EnrichController {

    private final EnrichmentService enrichmentService;

    public EnrichController(EnrichmentService enrichmentService) {
        this.enrichmentService = enrichmentService;
    }

    @PostMapping
    @ResponseStatus(HttpStatus.ACCEPTED)
    @Operation(summary = "Zlecenie wzbogacania (asynchroniczne)",
        description = "scope: SINGLE/SELECTED (z spotifyIds, max 100), MISSING (wg braków) "
            + "albo OUTDATED (utwory opisane starszym modelem/promptem — wyłącznie grupa AI); "
            + "fields: podzbiór METADATA/AUDIO/AI. Zlecenie ponad llm.max-tracks-per-job "
            + "kończy się 400 ENRICH_TOO_MANY_TRACKS — sprawdź wcześniej /api/enrich/estimate.")
    public Map<String, Long> startEnrichment(@Valid @RequestBody EnrichRequest request) {

        long executionId = enrichmentService.start(
            request.scope(), request.fields(), request.spotifyIdsOrEmpty());
        return Map.of("executionId", executionId);
    }

    @PostMapping("/estimate")
    @Operation(summary = "Ile utworów obejmie zlecenie i ile będzie kosztowało",
        description = "Nie uruchamia niczego. Koszt liczony tylko dla grupy AI — metadane "
            + "i cechy audio jadą z darmowych źródeł. Pusty koszt oznacza brak stawek "
            + "w konfiguracji (llm.cost.input-per-1m / llm.cost.output-per-1m), nie zero.")
    public EnrichmentEstimateResponse estimate(@Valid @RequestBody EnrichRequest request) {

        EnrichmentEstimate estimate = enrichmentService.estimate(
            request.scope(), request.fields(), request.spotifyIdsOrEmpty());
        return new EnrichmentEstimateResponse(
            estimate.trackCount(), estimate.aiTracks(), estimate.estimatedCost(),
            estimate.limit(), estimate.withinLimit());
    }

    @GetMapping("/jobs")
    @Operation(summary = "Ostatnie wykonania jobów wzbogacania")
    public List<EnrichJobResponse> listJobs(@RequestParam(defaultValue = "20") int limit) {

        return enrichmentService.listJobs(CatalogController.cappedSize(limit)).stream()
            .map(EnrichJobResponse::from)
            .toList();
    }

    @GetMapping("/jobs/{executionId}")
    @Operation(summary = "Status i postęp wykonania joba")
    public EnrichJobResponse jobStatus(@PathVariable long executionId) {
        return EnrichJobResponse.from(enrichmentService.status(executionId));
    }

    @GetMapping("/jobs/{executionId}/failures")
    @Operation(summary = "Utwory pominięte przez job razem z powodem",
        description = "Job nie przerywa się na pierwszym błędzie — przechodzi przez całą listę "
            + "i pomija to, co padło. Ta lista mówi, co dokładnie odpadło i dlaczego.")
    public List<EnrichFailureResponse> jobFailures(@PathVariable long executionId,
                                                   @RequestParam(defaultValue = "200") int limit) {

        return enrichmentService.failures(executionId, CatalogController.cappedSize(limit)).stream()
            .map(EnrichFailureResponse::from)
            .toList();
    }

    @PostMapping("/jobs/{executionId}/restart")
    @ResponseStatus(HttpStatus.ACCEPTED)
    @Operation(summary = "Restart nieudanego wykonania — dokańcza od checkpointu")
    public Map<String, Long> restartJob(@PathVariable long executionId) {
        return Map.of("executionId", enrichmentService.restart(executionId));
    }

    @GetMapping("/missing-count")
    @Operation(summary = "Liczba utworów z brakami per grupa pól")
    public MissingFieldsCount missingCount() {
        return enrichmentService.missingCount();
    }
}