# services/music-service/src/main/java/com/pgoogol/music/api/EnrichRequest.java
package com.pgoogol.music.api;

import com.pgoogol.music.enrichment.EnrichmentScope;
import com.pgoogol.music.enrichment.FieldGroup;
import jakarta.validation.constraints.NotEmpty;
import jakarta.validation.constraints.NotNull;

import java.util.List;
import java.util.Objects;
import java.util.Set;

/**
 * Zlecenie wzbogacania: zakres + grupy pól; {@code spotifyIds}
 * wymagane dla SINGLE/SELECTED.
 */
public record EnrichRequest(
    @NotNull EnrichmentScope scope,
    @NotEmpty Set<FieldGroup> fields,
    List<String> spotifyIds) {

    public List<String> spotifyIdsOrEmpty() {
        return Objects.requireNonNullElse(spotifyIds, List.of());
    }
}